+++ kroxylicious-kms-tls-support/src/main/java/io/kroxylicious/proxy/tls/TlsHttpClientConfigurator.java
@@ -7,14 +7,23 @@
 package io.kroxylicious.proxy.tls;
 
 import java.io.FileInputStream;
+import java.net.http.HttpClient;
+import java.net.http.HttpClient.Builder;
 import java.security.KeyStore;
+import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
-import java.security.cert.X509Certificate;
+import java.util.Arrays;
+import java.util.Objects;
 import java.util.Optional;
+import java.util.function.Function;
+import java.util.function.Predicate;
+import java.util.function.UnaryOperator;
+import java.util.stream.Collectors;
 
 import javax.net.ssl.KeyManager;
 import javax.net.ssl.KeyManagerFactory;
 import javax.net.ssl.SSLContext;
+import javax.net.ssl.SSLParameters;
 import javax.net.ssl.TrustManager;
 import javax.net.ssl.TrustManagerFactory;
 import javax.net.ssl.X509TrustManager;
@@ -23,6 +32,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import io.kroxylicious.proxy.config.secret.PasswordProvider;
+import io.kroxylicious.proxy.config.tls.AllowDeny;
 import io.kroxylicious.proxy.config.tls.InsecureTls;
 import io.kroxylicious.proxy.config.tls.KeyPair;
 import io.kroxylicious.proxy.config.tls.KeyProvider;
@@ -38,60 +48,62 @@ import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 
 /**
- * Encapsulates parameters for an TLS connection with AWS.
- *
- * @param tls tls configuration
- *
- * TODO ability to restrict by TLS protocol and cipher suite (<a href="https://github.com/kroxylicious/kroxylicious/issues/1006">#1006</a>)
+ * Responsible for applying TLS configuration to a {@link HttpClient.Builder}.
  */
-public record JdkTls(Tls tls) {
+public class TlsHttpClientConfigurator implements UnaryOperator<Builder> {
 
-    private static final Logger logger = LoggerFactory.getLogger(JdkTls.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(TlsHttpClientConfigurator.class);
 
-    public JdkTls {
-        if (tls != null && tls.key() != null) {
-            logger.warn("TLS key material is currently not supported by this client");
+    private static final SSLContext PLATFORM_SSL_CONTEXT;
+
+    static {
+        try {
+            PLATFORM_SSL_CONTEXT = SSLContext.getDefault();
+        }
+        catch (NoSuchAlgorithmException nsae) {
+            var e = new ExceptionInInitializerError("Failed to access default SSL context for platform");
+            e.initCause(nsae);
+            throw e;
         }
     }
 
-    public static final X509TrustManager INSECURE_TRUST_MANAGER = new X509TrustManager() {
+    private static final X509TrustManager INSECURE_TRUST_MANAGER = new InsecureTrustManager();
 
-        // suppressing sonar security warning as we are intentionally throwing security out the window
-        @SuppressWarnings("java:S4830")
-        @Override
-        public void checkClientTrusted(X509Certificate[] chain, String authType) {
-            // do nothing, the api is to throw if not trusted
+    private static final TrustManager[] INSECURE_TRUST_MANAGERS = { INSECURE_TRUST_MANAGER };
+    @Nullable
+    private final Tls tls;
+
+    /**
+     * Creates a TLS configurator.
+     *
+     * @param tls tls parameters
+     */
+    public TlsHttpClientConfigurator(@Nullable Tls tls) {
+        if (tls != null && tls.key() != null) {
+            LOGGER.warn("TLS key material is currently not supported by this client");
         }
+        this.tls = tls;
+    }
 
-        // suppressing sonar security warning as we are intentionally throwing security out the window
-        @SuppressWarnings("java:S4830")
-        @Override
-        public void checkServerTrusted(X509Certificate[] chain, String authType) {
-            // do nothing, the api is to throw if not trusted
-        }
-
-        @Override
-        public X509Certificate[] getAcceptedIssuers() {
-            return new X509Certificate[0];
-        }
-    };
-    public static final TrustManager[] INSECURE_TRUST_MANAGERS = { INSECURE_TRUST_MANAGER };
-
-    public SSLContext sslContext() {
+    private SSLContext sslContext() {
         try {
-            if (tls != null) {
+            if (tls == null || (tls.trust() == null && tls.key() == null)) {
+                return PLATFORM_SSL_CONTEXT;
+            }
+            else {
                 TrustManager[] trustManagers = null;
                 KeyManager[] keyManagers = null;
+
                 if (tls.trust() != null) {
                     trustManagers = getTrustManagers(tls.trust());
                 }
                 if (tls.key() != null) {
                     keyManagers = getKeyManagers(tls.key());
                 }
-                return getSslContext(trustManagers, keyManagers);
-            }
-            else {
-                return SSLContext.getDefault();
+
+                SSLContext context = SSLContext.getInstance("TLS");
+                context.init(keyManagers, trustManagers, new SecureRandom());
+                return context;
             }
         }
         catch (Exception e) {
@@ -99,12 +111,67 @@ public record JdkTls(Tls tls) {
         }
     }
 
+    private SSLParameters sslParameters() {
+        var defaultSslParameters = PLATFORM_SSL_CONTEXT.getDefaultSSLParameters();
+        if (tls == null || (tls.protocols() == null && tls.cipherSuites() == null)) {
+            return defaultSslParameters;
+        }
+
+        var supportedSSLParameters = PLATFORM_SSL_CONTEXT.getSupportedSSLParameters();
+
+        var protocols = applyRestriction("protocol", tls.protocols(), defaultSslParameters, supportedSSLParameters, SSLParameters::getProtocols);
+        var cipherSuites = applyRestriction("cipher suite", tls.cipherSuites(), defaultSslParameters, supportedSSLParameters, SSLParameters::getCipherSuites);
+
+        defaultSslParameters.setProtocols(protocols);
+        defaultSslParameters.setCipherSuites(cipherSuites);
+
+        return defaultSslParameters;
+    }
+
+    @NonNull
+    private String[] applyRestriction(String subject, AllowDeny<String> allowDeny, SSLParameters defaultSslParameters, SSLParameters supportedSSLParameters,
+                                      Function<SSLParameters, String[]> sslParametersAccessor) {
+        var result = Arrays.stream(sslParametersAccessor.apply(defaultSslParameters)).toList();
+        if (allowDeny != null) {
+            var supported = Arrays.stream(sslParametersAccessor.apply(supportedSSLParameters)).collect(Collectors.toSet());
+            var allowed = allowDeny.allowed();
+
+            if (allowed != null && !allowed.isEmpty()) {
+                allowed.stream()
+                        .filter(Predicate.not(supported::contains))
+                        .forEach(unsupported -> LOGGER.warn("Ignoring allowed {} '{}' as it is not recognized by this platform (supported: {})",
+                                subject, unsupported, supported));
+                result = allowed.stream()
+                        .filter(supported::contains)
+                        .toList();
+            }
+
+            var denied = allowDeny.denied();
+            if (denied != null) {
+                denied.stream()
+                        .filter(Predicate.not(supported::contains))
+                        .forEach(unsupported -> LOGGER.warn(
+                                "Ignoring denied {}} '{}' as it is not recognized by this platform (supported: {})",
+                                subject, unsupported, supported));
+                result = result.stream()
+                        .filter(Predicate.not(denied::contains))
+                        .toList();
+            }
+
+            if (result.isEmpty()) {
+                throw new SslConfigurationException(
+                        "The configuration you have in place has resulted in no %ss being available. Allowed: %s, Denied: %s".formatted(subject, allowed, denied));
+            }
+        }
+        return result.toArray(new String[]{});
+    }
+
     @VisibleForTesting
     static KeyManager[] getKeyManagers(KeyProvider key) {
         return key.accept(new KeyProviderVisitor<>() {
             @Override
             public KeyManager[] visit(KeyPair keyPair) {
-                throw new SslConfigurationException("KeyPair is not supported by vault KMS yet");
+                throw new SslConfigurationException("KeyPair is not supported by this client");
             }
 
             @Override
@@ -134,21 +201,6 @@ public record JdkTls(Tls tls) {
                 return Optional.ofNullable(value).map(PasswordProvider::getProvidedPassword).map(String::toCharArray).orElse(null);
             }
         });
-    }
-
-    @NonNull
-    private static SSLContext getSslContext(TrustManager[] trustManagers, KeyManager[] keyManagers) {
-        try {
-            if (trustManagers == null && keyManagers == null) {
-                return SSLContext.getDefault();
-            }
-            SSLContext context = SSLContext.getInstance("TLS");
-            context.init(keyManagers, trustManagers, new SecureRandom());
-            return context;
-        }
-        catch (Exception e) {
-            throw new SslConfigurationException(e);
-        }
     }
 
     @VisibleForTesting
@@ -199,5 +251,20 @@ public record JdkTls(Tls tls) {
                 }
             }
         });
+    }
+
+    /**
+     * Applies TLS configuration to the supplied {@link Builder}.  If there is no
+     * TLS configuration to apply, this operation is a no-op.
+     *
+     * @param builder HTTP client builder
+     * @return HTTP client builder
+     */
+    @Override
+    public Builder apply(@NonNull Builder builder) {
+        Objects.requireNonNull(builder);
+        builder.sslContext(sslContext())
+                .sslParameters(sslParameters());
+        return builder;
     }
 }